// src/infrastructure/grpc/generated/course/types/stats.ts
export interface CourseStats {
  totalCourses: number;
  draftCourses: number;
  publishedCourses: number;
  unPublishedCourses: number;
}

export interface MonthlyCoursesEnrollment {
  month: number;
  count: number;
}

export interface GetInstructorCourseRatingStatsRequest {
  courseId: string;
  instructorId: string;
}
// ...
export interface GetInstructorCoursesStatsRequest {
  instructorId: string;
}
// ...
export interface GetCoursesStatsResponse {
  success?: CourseStats | undefined;
  error?: Error | undefined;
}
// ...
export interface GetInstructorCoursesStatsResponse {
  success?: InstructorCoursesStats | undefined;
  error?: Error | undefined;
}

export interface GetEnrollmentTrendResponse {
  success?: EnrollmentTrendStats | undefined;
  error?: Error | undefined;
}

export interface GetInstructorCourseEnrollmentTrendResponse {
  success?: InstructorCourseEnrollmentTrend | undefined;
  error?: Error | undefined;
}

export interface GetInstructorCourseRatingStatsResponse {
  success?: InstructorCourseRatingStats | undefined;
  error?: Error | undefined;
}
// ...
export const GetInstructorCourseRatingStatsRequest: MessageFns<GetInstructorCourseRatingStatsRequest> =
  {
    encode(
      message: GetInstructorCourseRatingStatsRequest,
      writer: BinaryWriter = new BinaryWriter(),
    ): BinaryWriter {
      if (message.courseId !== "") {
        writer.uint32(10).string(message.courseId);
      }
      if (message.instructorId !== "") {
        writer.uint32(18).string(message.instructorId);
      }
      return writer;
    },

    decode(
      input: BinaryReader | Uint8Array,
      length?: number,
    ): GetInstructorCourseRatingStatsRequest {
      const reader =
        input instanceof BinaryReader ? input : new BinaryReader(input);
      const end = length === undefined ? reader.len : reader.pos + length;
      const message = createBaseGetInstructorCourseRatingStatsRequest();
      while (reader.pos < end) {
        const tag = reader.uint32();
        switch (tag >>> 3) {
          case 1: {
            if (tag !== 10) {
              break;
            }

            message.courseId = reader.string();
            continue;
          }
          case 2: {
            if (tag !== 18) {
              break;
            }

            message.instructorId = reader.string();
            continue;
          }
        }
        if ((tag & 7) === 4 || tag === 0) {
          break;
        }
        reader.skip(tag & 7);
      }
      return message;
    },

    fromJSON(object: any): GetInstructorCourseRatingStatsRequest {
      return {
        courseId: isSet(object.courseId)
          ? globalThis.String(object.courseId)
          : isSet(object.course_id)
            ? globalThis.String(object.course_id)
            : "",
        instructorId: isSet(object.instructorId)
          ? globalThis.String(object.instructorId)
          : isSet(object.instructor_id)
            ? globalThis.String(object.instructor_id)
            : "",
      };
    },

    toJSON(message: GetInstructorCourseRatingStatsRequest): unknown {
      const obj: any = {};
      if (message.courseId !== "") {
        obj.courseId = message.courseId;
      }
      if (message.instructorId !== "") {
        obj.instructorId = message.instructorId;
      }
      return obj;
    },

    create<
      I extends Exact<DeepPartial<GetInstructorCourseRatingStatsRequest>, I>,
    >(base?: I): GetInstructorCourseRatingStatsRequest {
      return GetInstructorCourseRatingStatsRequest.fromPartial(
        base ?? ({} as any),
      );
    },
    fromPartial<
      I extends Exact<DeepPartial<GetInstructorCourseRatingStatsRequest>, I>,
    >(object: I): GetInstructorCourseRatingStatsRequest {
      const message = createBaseGetInstructorCourseRatingStatsRequest();
      message.courseId = object.courseId ?? "";
      message.instructorId = object.instructorId ?? "";
      return message;
    },
  };
// ...
export const GetInstructorCoursesStatsRequest: MessageFns<GetInstructorCoursesStatsRequest> =
  {
    encode(
      message: GetInstructorCoursesStatsRequest,
      writer: BinaryWriter = new BinaryWriter(),
    ): BinaryWriter {
      if (message.instructorId !== "") {
        writer.uint32(10).string(message.instructorId);
      }
      return writer;
    },

    decode(
      input: BinaryReader | Uint8Array,
      length?: number,
    ): GetInstructorCoursesStatsRequest {
      const reader =
        input instanceof BinaryReader ? input : new BinaryReader(input);
      const end = length === undefined ? reader.len : reader.pos + length;
      const message = createBaseGetInstructorCoursesStatsRequest();
      while (reader.pos < end) {
        const tag = reader.uint32();
        switch (tag >>> 3) {
          case 1: {
            if (tag !== 10) {
              break;
            }

            message.instructorId = reader.string();
            continue;
          }
        }
        if ((tag & 7) === 4 || tag === 0) {
          break;
        }
        reader.skip(tag & 7);
      }
      return message;
    },

    fromJSON(object: any): GetInstructorCoursesStatsRequest {
      return {
        instructorId: isSet(object.instructorId)
          ? globalThis.String(object.instructorId)
          : isSet(object.instructor_id)
            ? globalThis.String(object.instructor_id)
            : "",
      };
    },

    toJSON(message: GetInstructorCoursesStatsRequest): unknown {
      const obj: any = {};
      if (message.instructorId !== "") {
        obj.instructorId = message.instructorId;
      }
      return obj;
    },

    create<I extends Exact<DeepPartial<GetInstructorCoursesStatsRequest>, I>>(
      base?: I,
    ): GetInstructorCoursesStatsRequest {
      return GetInstructorCoursesStatsRequest.fromPartial(base ?? ({} as any));
    },
    fromPartial<
      I extends Exact<DeepPartial<GetInstructorCoursesStatsRequest>, I>,
    >(object: I): GetInstructorCoursesStatsRequest {
      const message = createBaseGetInstructorCoursesStatsRequest();
      message.instructorId = object.instructorId ?? "";
      return message;
    },
  };
// ...
export const GetCoursesStatsResponse: MessageFns<GetCoursesStatsResponse> = {
  encode(
    message: GetCoursesStatsResponse,
    writer: BinaryWriter = new BinaryWriter(),
  ): BinaryWriter {
    if (message.success !== undefined) {
      CourseStats.encode(message.success, writer.uint32(10).fork()).join();
    }
    if (message.error !== undefined) {
      Error.encode(message.error, writer.uint32(18).fork()).join();
    }
    return writer;
  },

  decode(
    input: BinaryReader | Uint8Array,
    length?: number,
  ): GetCoursesStatsResponse {
    const reader =
      input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseGetCoursesStatsResponse();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.success = CourseStats.decode(reader, reader.uint32());
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.error = Error.decode(reader, reader.uint32());
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): GetCoursesStatsResponse {
    return {
      success: isSet(object.success)
        ? CourseStats.fromJSON(object.success)
        : undefined,
      error: isSet(object.error) ? Error.fromJSON(object.error) : undefined,
    };
  },

  toJSON(message: GetCoursesStatsResponse): unknown {
    const obj: any = {};
    if (message.success !== undefined) {
      obj.success = CourseStats.toJSON(message.success);
    }
    if (message.error !== undefined) {
      obj.error = Error.toJSON(message.error);
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<GetCoursesStatsResponse>, I>>(
    base?: I,
  ): GetCoursesStatsResponse {
    return GetCoursesStatsResponse.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<GetCoursesStatsResponse>, I>>(
    object: I,
  ): GetCoursesStatsResponse {
    const message = createBaseGetCoursesStatsResponse();
    message.success =
      object.success !== undefined && object.success !== null
        ? CourseStats.fromPartial(object.success)
        : undefined;
    message.error =
      object.error !== undefined && object.error !== null
        ? Error.fromPartial(object.error)
        : undefined;
    return message;
  },
};
// ...
export const GetInstructorCourseRatingStatsResponse: MessageFns<GetInstructorCourseRatingStatsResponse> =
  {
    encode(
      message: GetInstructorCourseRatingStatsResponse,
      writer: BinaryWriter = new BinaryWriter(),
    ): BinaryWriter {
      if (message.success !== undefined) {
        InstructorCourseRatingStats.encode(
          message.success,
          writer.uint32(10).fork(),
        ).join();
      }
      if (message.error !== undefined) {
        Error.encode(message.error, writer.uint32(18).fork()).join();
      }
      return writer;
    },

    decode(
      input: BinaryReader | Uint8Array,
      length?: number,
    ): GetInstructorCourseRatingStatsResponse {
      const reader =
        input instanceof BinaryReader ? input : new BinaryReader(input);
      const end = length === undefined ? reader.len : reader.pos + length;
      const message = createBaseGetInstructorCourseRatingStatsResponse();
      while (reader.pos < end) {
        const tag = reader.uint32();
        switch (tag >>> 3) {
          case 1: {
            if (tag !== 10) {
              break;
            }

            message.success = InstructorCourseRatingStats.decode(
              reader,
              reader.uint32(),
            );
            continue;
          }
          case 2: {
            if (tag !== 18) {
              break;
            }

            message.error = Error.decode(reader, reader.uint32());
            continue;
          }
        }
        if ((tag & 7) === 4 || tag === 0) {
          break;
        }
        reader.skip(tag & 7);
      }
      return message;
    },

    fromJSON(object: any): GetInstructorCourseRatingStatsResponse {
      return {
        success: isSet(object.success)
          ? InstructorCourseRatingStats.fromJSON(object.success)
          : undefined,
        error: isSet(object.error) ? Error.fromJSON(object.error) : undefined,
      };
    },

    toJSON(message: GetInstructorCourseRatingStatsResponse): unknown {
      const obj: any = {};
      if (message.success !== undefined) {
        obj.success = InstructorCourseRatingStats.toJSON(message.success);
      }
      if (message.error !== undefined) {
        obj.error = Error.toJSON(message.error);
      }
      return obj;
    },

    create<
      I extends Exact<DeepPartial<GetInstructorCourseRatingStatsResponse>, I>,
    >(base?: I): GetInstructorCourseRatingStatsResponse {
      return GetInstructorCourseRatingStatsResponse.fromPartial(
        base ?? ({} as any),
      );
    },
    fromPartial<
      I extends Exact<DeepPartial<GetInstructorCourseRatingStatsResponse>, I>,
    >(object: I): GetInstructorCourseRatingStatsResponse {
      const message = createBaseGetInstructorCourseRatingStatsResponse();
      message.success =
        object.success !== undefined && object.success !== null
          ? InstructorCourseRatingStats.fromPartial(object.success)
          : undefined;
      message.error =
        object.error !== undefined && object.error !== null
          ? Error.fromPartial(object.error)
          : undefined;
      return message;
    },
  };
// ...
export type Exact<P, I extends P> = P extends Builtin
  ? P
  : P & { [K in keyof P]: Exact<P[K], I[K]> } & {
      [K in Exclude<keyof I, KeysOfUnion<P>>]: never;
    };

function isSet(value: any): boolean {
  return value !== null && value !== undefined;
}

export interface MessageFns<T> {
  encode(message: T, writer?: BinaryWriter): BinaryWriter;
  decode(input: BinaryReader | Uint8Array, length?: number): T;
  fromJSON(object: any): T;
  toJSON(message: T): unknown;
  create<I extends Exact<DeepPartial<T>, I>>(base?: I): T;
  fromPartial<I extends Exact<DeepPartial<T>, I>>(object: I): T;
}